+++ autode/methods.py
@@ -5,17 +5,18 @@ from autode.wrappers.MOPAC import MOPAC
 from autode.wrappers.NWChem import NWChem
 from autode.wrappers.ORCA import ORCA
 from autode.wrappers.XTB import XTB
+from autode.wrappers.QChem import QChem
 from autode.config import Config
 from autode.exceptions import MethodUnavailable
 
 """
 Functions to get the high and low level electronic structure methods to use 
 for example high-level methods would be orca and Gaussian09 which can perform 
-DFT/WF theory calculations, low level methods are for example xtb and mopac 
-which are non ab-initio methods and are therefore considerably faster
+DFT/WF theory calculations, low level methods are, for example, xtb and mopac 
+which are fast non ab-initio methods
 """
 
-high_level_method_names = ['orca', 'g09', 'g16', 'nwchem']
+high_level_method_names = ['orca', 'g09', 'g16', 'nwchem', 'qchem']
 low_level_method_names = ['xtb', 'mopac']
 
 
@@ -25,7 +26,7 @@ def get_hmethod() -> ElectronicStructureMethod:
     Returns:
         (autode.wrappers.base.ElectronicStructureMethod): High-level method
     """
-    h_methods = [ORCA(), G09(), NWChem(), G16()]
+    h_methods = [ORCA(), G09(), NWChem(), G16(), QChem()]
 
     if Config.hcode is not None:
         return get_defined_method(name=Config.hcode, possibilities=h_methods)
@@ -39,7 +40,7 @@ def get_lmethod() -> ElectronicStructureMethod:
     Returns:
         (autode.wrappers.base.ElectronicStructureMethod): Low-level method
     """
-    all_methods = [XTB(), MOPAC(), ORCA(), G16(), G09(), NWChem()]
+    all_methods = [XTB(), MOPAC(), ORCA(), G16(), G09(), NWChem(), QChem()]
 
     if Config.lcode is not None:
         return get_defined_method(name=Config.lcode, possibilities=all_methods)