+++ autode/wrappers/dispersion.py
@@ -3,4 +3,5 @@ from autode.wrappers.keywords import DispersionCorrection
 d3bj = DispersionCorrection(name='d3bj',
                             doi_list=['10.1063/1.3382344', '10.1002/jcc.21759'],
                             g09='GD3BJ',
-                            orca='D3BJ')
+                            orca='D3BJ',
+                            qchem='D3_BJ')